+++ SDRSharp.SatnogsTracker/Properties/AssemblyInfo.cs
@@ -43,3 +43,4 @@ using System.Runtime.InteropServices;
 
 [assembly: AssemblyVersion("1.5.*")]
 [assembly: AssemblyFileVersion(version: "1.5.3")]
+