+++ SDRSharp.SatnogsTracker/SatnogsTrackerPlugin.cs
@@ -38,6 +38,10 @@ using Zeptomoby.OrbitTools;
 using SDRSharp.PanView;
 using System.Drawing;
 using Newtonsoft.Json;
+using System.Net.Sockets;
+using System.Runtime.InteropServices;
+using NAudio.Wave;
+using NAudio.Wave.Compression;
 
 namespace SDRSharp.SatnogsTracker
 {
@@ -50,7 +54,11 @@ namespace SDRSharp.SatnogsTracker
         public Boolean CalculateSatVisibilityRunning;
         private readonly RecordingIQProcessor _iqObserver = new RecordingIQProcessor();
         private readonly RecordingAudioProcessor _audioProcessor = new RecordingAudioProcessor();
+        private readonly RecordingAudioProcessor _UDPaudioProcessor = new RecordingAudioProcessor();
         private SimpleRecorder _audioRecorder;
+        private SimpleStreamer _UDPaudioStreamer;
+
+
         private SimpleRecorder _basebandRecorder;
         private readonly WavSampleFormat _wavSampleFormat = WavSampleFormat.PCM16;
         private string _SatElevation;
@@ -74,11 +82,14 @@ namespace SDRSharp.SatnogsTracker
 
             /// Setup Audio Recording
             _audioProcessor.Enabled = false;
+            _UDPaudioProcessor.Enabled = false;
             _iqObserver.Enabled = false;
             control_.RegisterStreamHook(_iqObserver, ProcessorType.RawIQ);
-            control_.RegisterStreamHook(_audioProcessor, ProcessorType.DemodulatorOutput);
-
+            control_.RegisterStreamHook(_audioProcessor, ProcessorType.FilteredAudioOutput);
+            control_.RegisterStreamHook(_UDPaudioProcessor, ProcessorType.FilteredAudioOutput);
+            Console.WriteLine(_audioProcessor.SampleRate);
             _audioRecorder = new SimpleRecorder(_audioProcessor);
+            _UDPaudioStreamer = new SimpleStreamer(_UDPaudioProcessor,"127.0.0.1",7355);
             _basebandRecorder = new SimpleRecorder(_iqObserver);
 
             //Instanciate all needed objects
@@ -149,6 +160,8 @@ namespace SDRSharp.SatnogsTracker
 
             satpc32Server.SatRecordAFChanged += _controlpanel.SatPC32ServerRecordAFChanged;
             satpc32Server.SatRecordAFChanged += SDRSharp_AFRecorderChanged;
+
+            _controlpanel.StartRecordingAF += SDRSharp_AFRecorderChanged;
             #endregion
 
             #region Load and Setup Satellite and Transmitter data to Memory
@@ -158,6 +171,8 @@ namespace SDRSharp.SatnogsTracker
             Console.WriteLine("Tracking Initiated");
             //Start Background Doppler Calculations
             ThreadPool.QueueUserWorkItem(CalculateSatVisibility);
+
+
         }
 
         private void  UpdateTleData(Boolean alwaysdownload)
@@ -182,6 +197,7 @@ namespace SDRSharp.SatnogsTracker
                 dlg.HamSiteChanged += _controlpanel.ControlPanel_HomeSite_Changed;
                 dlg.UpdateTle += UpdateTleData;
                 UpdateStatus += dlg.UpdateStatus;
+                dlg.UpdateIP +=_UDPaudioStreamer.UpdateIP;
                 dlg.Show();
             }
         }
@@ -229,6 +245,7 @@ namespace SDRSharp.SatnogsTracker
             CalculateSatVisibilityRunning = false;
             StopBaseRecorder();
             StopAFRecorder();
+            //StopUDPStreamer();
             satpc32Server_?.Abort();
             LogFile.Close();
         }
@@ -309,8 +326,8 @@ namespace SDRSharp.SatnogsTracker
             //
             //http://www.celestrak.com/NORAD/elements/amateur.txt
             // DK3WN has a better amateur.txt file
-            //
-            GetFile("http://www.dk3wn.info/tle/amateur.txt", "amateur.txt", alwaysdownload);
+            //http://www.dk3wn.info/tle/amateur.txt
+            GetFile("http://www.celestrak.com/NORAD/elements/amateur.txt", "amateur.txt", alwaysdownload);
             LoadTxtKeps("amateur.txt");
 
             GetFile("http://www.celestrak.com/NORAD/elements/active.txt", "active.txt", alwaysdownload);
@@ -347,6 +364,7 @@ namespace SDRSharp.SatnogsTracker
                     {
                         if (_basebandRecorder.IsRecording) StopBaseRecorder();
                         if (_audioRecorder.IsRecording) StopAFRecorder();
+                        if (_UDPaudioStreamer.IsStreaming) StopUDPStreamer();
                     }
                 }
             }
@@ -447,6 +465,9 @@ namespace SDRSharp.SatnogsTracker
                     {
                         PrepareAFRecorder();
                         _audioRecorder.StartRecording();
+
+                        PrepareUDPStreamer();
+                        _UDPaudioStreamer.StartStreaming();
                     }
                 }
                 catch
@@ -455,6 +476,11 @@ namespace SDRSharp.SatnogsTracker
                     MessageBox.Show("Unable to Start AF Recording", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
                     return;
                 }
+            }
+
+            if (!RecordAF && _audioRecorder.IsRecording)
+            {
+                _audioRecorder.StopRecording();
             }
         }
 
@@ -469,10 +495,25 @@ namespace SDRSharp.SatnogsTracker
         private void PrepareAFRecorder()
         {
             DateTime startTime = DateTime.UtcNow;
-            _audioRecorder.SampleRate = _audioProcessor.SampleRate;
-            String AudioRecordingName = startTime.ToString(@"yyyy-MM-ddTHH-mm.ffffff") + "_" + SatelliteName + "_" + SatelliteID + "_AF.wav";
+            String AudioRecordingName = "";
+            //_audioProcessor.SampleRate = 48000;
+            _audioRecorder.SampleRate = _audioProcessor.SampleRate; 
+            if ((SatelliteName==null) || (SatelliteID==null))
+            {
+                AudioRecordingName = startTime.ToString(@"yyyy-MM-ddTHH-mm.ffffff") + "CURRENT_FREQ__AF.wav";
+            } else
+                AudioRecordingName = startTime.ToString(@"yyyy-MM-ddTHH-mm.ffffff") + "_" + SatelliteName + "_" + SatelliteID + "_AF.wav";
             _audioRecorder.FileName = RecordingLocation() + "\\" + AudioRecordingName;
             _audioRecorder.Format = _wavSampleFormat;
+        }
+
+        private void PrepareUDPStreamer()
+        {
+            DateTime startTime = DateTime.UtcNow;
+            _UDPaudioStreamer.SampleRate = _audioProcessor.SampleRate;
+            String AudioRecordingName = startTime.ToString(@"yyyy-MM-ddTHH-mm.ffffff") + "_" + SatelliteName + "_" + SatelliteID + "_STREAM.wav";
+            _UDPaudioStreamer.FileName = RecordingLocation() + "\\" + AudioRecordingName;
+            _UDPaudioStreamer.Format = _wavSampleFormat;
         }
         private void StopBaseRecorder()
         {
@@ -481,6 +522,11 @@ namespace SDRSharp.SatnogsTracker
         private void StopAFRecorder()
         {
             if (_audioRecorder.IsRecording) _audioRecorder.StopRecording();
+        }
+
+        private void StopUDPStreamer()
+        {
+            if (_UDPaudioStreamer.IsStreaming) _UDPaudioStreamer.StopStreaming();
         }
 
         private void SDRSharp_WaterFallCustomPaint(object sender, CustomPaintEventArgs e)
@@ -1895,5 +1941,535 @@ namespace SDRSharp.SatnogsTracker
     }
     #endregion
 
-    
+    public enum RecordingMode
+    {
+        Baseband,
+        Audio
+    }
+
+    public unsafe class SimpleStreamer : IDisposable
+    {
+        private const int DefaultAudioGain = 30;
+        private const long MaxStreamLength = int.MaxValue;
+        private static readonly int _bufferCount = Utils.GetIntSetting("RecordingBufferCount", 8);
+        private readonly float _audioGain = (float)Math.Pow(DefaultAudioGain / 10.0, 10);
+
+        private readonly SharpEvent _bufferEvent = new SharpEvent(false);
+
+        private readonly UnsafeBuffer[] _circularBuffers = new UnsafeBuffer[_bufferCount];
+        private readonly Complex*[] _complexCircularBufferPtrs = new Complex*[_bufferCount];
+        private readonly float*[] _floatCircularBufferPtrs = new float*[_bufferCount];
+
+        private int _circularBufferTail;
+        private int _circularBufferHead;
+        private int _circularBufferLength;
+        private volatile int _circularBufferUsedCount;
+
+        private long _skippedBuffersCount;
+        private bool _diskWriterRunning;
+        private string _fileName;
+        private double _sampleRate;
+
+        private WavSampleFormat _wavSampleFormat;
+        private SimpleWavWriter _wavWriter;
+        private Thread _diskWriter;
+
+        private readonly RecordingMode _recordingMode;
+        //private readonly RecordingIQObserver _iQObserver;
+        private readonly RecordingAudioProcessor _audioProcessor;
+        private byte[] msgBuffer = new byte[40000];
+        private byte[] _outputBuffer = null;
+        private BinaryWriter _outputStream;
+        private long _length;
+        private Boolean _isStreamFull;
+        public bool IsStreaming
+        {
+            get { return _diskWriterRunning; }
+        }
+
+        public bool IsStreamFull
+        {
+            get { return _wavWriter == null ? false : _wavWriter.IsStreamFull; }
+        }
+
+        public long BytesWritten
+        {
+            get { return _wavWriter == null ? 0L : _wavWriter.Length; }
+        }
+
+        public long SkippedBuffers
+        {
+            get { return _wavWriter == null ? 0L : _skippedBuffersCount; }
+        }
+
+        public RecordingMode Mode
+        {
+            get { return _recordingMode; }
+        }
+
+        public WavSampleFormat Format
+        {
+            get { return _wavSampleFormat; }
+            set
+            {
+                if (_diskWriterRunning)
+                {
+                    throw new ArgumentException("Format cannot be set while recording");
+                }
+                _wavSampleFormat = value;
+            }
+        }
+
+        public double SampleRate
+        {
+            get { return _sampleRate; }
+            set
+            {
+                if (_diskWriterRunning)
+                {
+                    throw new ArgumentException("SampleRate cannot be set while recording");
+                }
+
+                _sampleRate = value;
+            }
+        }
+
+        public string FileName
+        {
+            get { return _fileName; }
+            set
+            {
+                if (_diskWriterRunning)
+                {
+                    throw new ArgumentException("FileName cannot be set while recording");
+                }
+                _fileName = value;
+            }
+        }
+
+
+    #region Initialization and Termination
+    private UdpClient _udpClient;
+        private IPEndPoint _udpEP;
+        private int _port;
+
+        public SimpleStreamer(RecordingAudioProcessor audioProcessor,String Host, int Port)
+        {
+            _audioProcessor = audioProcessor;
+            _recordingMode = RecordingMode.Audio;
+            _udpClient = new UdpClient();
+            _port = Port;
+            _udpEP = new IPEndPoint(IPAddress.Parse(Host), Port); // endpoint where server is listening
+            //_udpClient.Connect(_udpEP);
+        }
+
+        ~SimpleStreamer()
+        {
+            Dispose();
+        }
+
+        public void Dispose()
+        {
+            FreeBuffers();
+        }
+
+        #endregion
+
+        #region IQ Event Handler
+
+        public void IQSamplesIn(Complex* buffer, int length)
+        {
+
+            #region Buffers
+
+            if (_circularBufferLength != length)
+            {
+                FreeBuffers();
+                CreateBuffers(length);
+
+                _circularBufferTail = 0;
+                _circularBufferHead = 0;
+            }
+
+            #endregion
+
+            if (_circularBufferUsedCount == _bufferCount)
+            {
+                _skippedBuffersCount++;
+                return;
+            }
+
+            Utils.Memcpy(_complexCircularBufferPtrs[_circularBufferHead], buffer, length * sizeof(Complex));
+            _circularBufferHead++;
+            _circularBufferHead &= (_bufferCount - 1);
+            _circularBufferUsedCount++;
+            _bufferEvent.Set();
+        }
+
+        #endregion
+
+        #region Audio Event / Scaling
+
+        public void AudioSamplesIn(float* audio, int length)
+        {
+            #region Buffers
+
+            var sampleCount = length / 2;
+            if (_circularBufferLength != sampleCount)
+            {
+                FreeBuffers();
+                CreateBuffers(sampleCount);
+
+                _circularBufferTail = 0;
+                _circularBufferHead = 0;
+            }
+
+            #endregion
+
+            if (_circularBufferUsedCount == _bufferCount)
+            {
+                _skippedBuffersCount++;
+                return;
+            }
+
+            Utils.Memcpy(_floatCircularBufferPtrs[_circularBufferHead], audio, length * sizeof(float));
+            _circularBufferHead++;
+            _circularBufferHead &= (_bufferCount - 1);
+            _circularBufferUsedCount++;
+            _bufferEvent.Set();
+        }
+
+        public void ScaleAudio(float* audio, int length)
+        {
+            for (var i = 0; i < length; i++)
+            {
+                audio[i] *= _audioGain;
+            }
+        }
+
+        #endregion
+
+        #region Worker Thread
+        public AcmStream resampleStream;
+        private void DiskWriterThread()
+        {
+            if (_recordingMode == RecordingMode.Audio)
+            {
+                _audioProcessor.AudioReady += AudioSamplesIn;
+                _audioProcessor.Enabled = true;
+            }
+            int input_rate = (int)_sampleRate;
+            resampleStream = new AcmStream(new WaveFormat(input_rate, 16, 1), new WaveFormat(48000, 16, 1));
+            
+            while (_diskWriterRunning)// && !_wavWriter.IsStreamFull)
+            {
+                if (_circularBufferTail == _circularBufferHead)
+                {
+                    _bufferEvent.WaitOne();
+                }
+
+                if (_diskWriterRunning && _circularBufferTail != _circularBufferHead)
+                {
+                    if (_recordingMode == RecordingMode.Audio)
+                    {
+                       ScaleAudio(_floatCircularBufferPtrs[_circularBufferTail], _circularBuffers[_circularBufferTail].Length * 2);
+                    }
+                    //byte []bytes = new byte[_circularBuffers[_circularBufferTail].Length * 2];
+
+                    /*
+                    int counter = 0;
+                    for (int i = 0; i < _circularBuffers[_circularBufferTail].Length * 2; i++)
+                    {
+                        byte[] bytes = BitConverter.GetBytes(_floatCircularBufferPtrs[_circularBufferTail][i]);// _floatCircularBufferPtrs[_circularBufferTail][i];
+                        System.Buffer.BlockCopy(bytes, 0, msgBuffer, counter, bytes.Length);
+                        counter += bytes.Length;
+                        if (counter>3000)
+                        {
+                            try
+                            {
+                                //_udpClient.Send(sendBytes, sendBytes.Length, _udpEP);
+                                _udpClient.Send(msgBuffer, counter, _udpEP);
+
+                            }
+                            catch (Exception e)
+                            {
+                                Console.WriteLine(e.ToString());
+                            }
+                            counter = 0;
+                        }
+                    }
+                        //bytes[i] = (byte)_floatCircularBufferPtrs[_circularBufferTail][i];
+
+                    */
+                        
+                     Write(_floatCircularBufferPtrs[_circularBufferTail], _circularBuffers[_circularBufferTail].Length);
+
+                    _circularBufferUsedCount--;
+                    _circularBufferTail++;
+                    _circularBufferTail &= (_bufferCount - 1);
+                }
+            }
+
+            while (_circularBufferTail != _circularBufferHead)
+            {
+                if (_floatCircularBufferPtrs[_circularBufferTail] != null)
+                {
+                    //_wavWriter.Write(_floatCircularBufferPtrs[_circularBufferTail], _circularBuffers[_circularBufferTail].Length);
+                }
+                _circularBufferTail++;
+                _circularBufferTail &= (_bufferCount - 1);
+            }
+
+            if (_recordingMode == RecordingMode.Audio)
+            { 
+                _audioProcessor.Enabled = false;
+                _audioProcessor.AudioReady -= AudioSamplesIn;
+            }
+
+            _diskWriterRunning = false;
+        }
+
+        public void Write(float* data, int length)
+        {
+
+            if (_udpClient != null)
+            {
+
+                //if (_isStreamFull)
+                //{
+                //    throw new InvalidOperationException("Stream full");
+                //}
+
+                switch (_wavSampleFormat)
+                {
+                    case WavSampleFormat.PCM8:
+                        WritePCM8(data, length);
+                        break;
+
+                    case WavSampleFormat.PCM16:
+                        WritePCM16(data, length);
+                        break;
+
+                    case WavSampleFormat.Float32:
+                        WriteFloat(data, length);
+                        break;
+                }
+
+                return;
+            }
+
+            throw new InvalidOperationException("Stream not open");
+        }
+        private void WritePCM8(float* data, int length)
+        {
+
+            #region Buffer
+
+            if (_outputBuffer == null || _outputBuffer.Length != length)
+            {
+                _outputBuffer = null;
+                _outputBuffer = new byte[length];// * 2];
+            }
+
+            #endregion
+
+            var ptr = data;
+
+            for (var i = 0; i < length; i++)
+            {
+                _outputBuffer[i]=(byte)((*ptr++ * 127.0f) + 128);
+                /*
+                _outputBuffer[(i * 2)] = (byte)((*ptr++ * 127.0f) + 128);
+                _outputBuffer[(i * 2) + 1] = (byte)((*ptr++ * 127.0f) + 128);
+                */
+            }
+
+            WriteStream(_outputBuffer);
+        }
+
+        private void WritePCM16(float* data, int length)
+        {
+            if (_outputBuffer == null || _outputBuffer.Length != (length * sizeof(Int16)))
+            {
+                _outputBuffer = null;
+                _outputBuffer = new byte[length * sizeof(Int16)];
+            }
+            var ptr = data;           
+            for (var i = 0; i < length; i++)
+            {
+                var leftChannel = (Int16)(*ptr++ * 32767.0f);
+                var rightChannel = (Int16)(*ptr++ * 32767.0f);
+                _outputBuffer[(i * 2)] = (byte)(leftChannel & 0x00ff);
+                _outputBuffer[(i * 2) + 1] = (byte)(leftChannel >> 8);
+            }
+            int buffer_length = length * sizeof(Int16);
+            System.Buffer.BlockCopy(_outputBuffer,0,resampleStream.SourceBuffer,0,buffer_length);
+            int sourceBytesConverted = 0;
+            
+            var convertedBytes = resampleStream.Convert(buffer_length, out sourceBytesConverted);
+            if (sourceBytesConverted != buffer_length)
+            {
+                Console.WriteLine("We didn't convert everything {0} bytes in, {1} bytes converted");
+            }
+            var converted = new byte[convertedBytes];
+            System.Buffer.BlockCopy(resampleStream.DestBuffer, 0, converted, 0, convertedBytes);
+            int counter = 0;
+            byte[] packet = new byte[1471];
+            for (var i=0; i< convertedBytes; i++)
+            {
+                
+                if (counter == 1471)
+                {
+                    _udpClient.Send(packet, counter, _udpEP);
+                    counter = 0;
+                }
+                packet[counter] = converted[i];
+                counter++;
+            }
+            if (counter > 0)
+            {
+                _udpClient.Send(packet, counter, _udpEP);
+            }
+        }
+
+        private void WriteFloat(float* data, int length)
+        {
+
+            #region Buffer
+
+            if (_outputBuffer == null || _outputBuffer.Length != (length * sizeof(float) * 2))
+            {
+                _outputBuffer = null;
+                _outputBuffer = new byte[length * sizeof(float) * 2];
+            }
+
+            #endregion
+
+            Marshal.Copy((IntPtr)data, _outputBuffer, 0, _outputBuffer.Length);
+
+            WriteStream(_outputBuffer);
+        }
+
+        private void WriteStream(byte[] data)
+        {
+            _udpClient.Send(data, data.Length, _udpEP);
+            /*
+            if (_outputStream != null)
+            {
+                var toWrite = (int)Math.Min(MaxStreamLength - _outputStream.BaseStream.Length, data.Length);
+
+                _outputStream.Write(data, 0, toWrite);
+
+                _length += toWrite;
+                UpdateLength();
+
+                _isStreamFull = _outputStream.BaseStream.Length >= MaxStreamLength;
+            }
+            */
+        }
+        public void InsertDataIntoCircularBuffer(byte[] data)
+        {
+
+        }
+        public void SendingPackets()
+        {
+
+        }
+        private void UpdateLength()
+        {
+            if (_outputStream != null)
+            {
+                /*
+                _outputStream.Seek((int)_fileSizeOffs, SeekOrigin.Begin);
+                _outputStream.Write((UInt32)(_outputStream.BaseStream.Length - 8));
+                _outputStream.Seek((int)_dataSizeOffs, SeekOrigin.Begin);
+                _outputStream.Write((UInt32)(_length));
+                _outputStream.BaseStream.Seek(0, SeekOrigin.End);
+                */
+            }
+        }
+        private void Flush()
+        {
+            if (_wavWriter != null)
+            {
+                //_wavWriter.Close();
+            }
+        }
+
+        private void CreateBuffers(int size)
+        {
+            for (var i = 0; i < _bufferCount; i++)
+            {
+                _circularBuffers[i] = UnsafeBuffer.Create(size, sizeof(Complex));
+                _complexCircularBufferPtrs[i] = (Complex*)_circularBuffers[i];
+                _floatCircularBufferPtrs[i] = (float*)_circularBuffers[i];
+            }
+
+            _circularBufferLength = size;
+        }
+
+        private void FreeBuffers()
+        {
+            _circularBufferLength = 0;
+            for (var i = 0; i < _bufferCount; i++)
+            {
+                if (_circularBuffers[i] != null)
+                {
+                    _circularBuffers[i].Dispose();
+                    _circularBuffers[i] = null;
+                    _complexCircularBufferPtrs[i] = null;
+                    _floatCircularBufferPtrs[i] = null;
+                }
+            }
+        }
+
+        #endregion
+
+        #region Public Methods
+
+        public void StartStreaming()
+        {
+            if (_diskWriter == null)
+            {
+                _circularBufferHead = 0;
+                _circularBufferTail = 0;
+
+                _skippedBuffersCount = 0;
+
+                _bufferEvent.Reset();
+
+                //_wavWriter = new SimpleWavWriter(_fileName, _wavSampleFormat, (uint)_sampleRate);
+                //_wavWriter.Open();
+
+                _diskWriter = new Thread(DiskWriterThread);
+
+                _diskWriterRunning = true;
+                _diskWriter.Start();
+            }
+        }
+
+        public void StopStreaming()
+        {
+            _diskWriterRunning = false;
+
+            if (_diskWriter != null)
+            {
+                _bufferEvent.Set();
+                _diskWriter.Join();
+            }
+
+            Flush();
+            FreeBuffers();
+
+            _diskWriter = null;
+            _wavWriter = null;
+        }
+
+        internal void UpdateIP(string NewIP)
+        {
+            _udpEP = new IPEndPoint(IPAddress.Parse(NewIP), _port);
+        }
+
+        #endregion
+    }
 }